requirements.txt
# Core dependencies
streamlit==1.28.2
torch==2.1.0
torchvision==0.16.0

# Image processing
Pillow==10.1.0
numpy==1.26.0
pandas==2.1.0
scikit-image==0.22.0
scipy==1.11.0
opencv-python-headless==4.8.1

# Data visualization
plotly==5.17.0
matplotlib==3.8.0

# MLops and monitoring
mlflow==2.9.0
psutil==5.9.5
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pyyaml==6.0.1
watchdog==3.0.0

# Development
black==23.11.0
flake8==6.1.0
pytest==7.4.0